Add proper latency reporting for the Android Java API
This change introduces output latency calculation based on timestamps
for the Android Java API.  Because timestamps for samples are only
available after the first sample of the stream has reached the audio
output, the implementation also adds a callback function that lets the
client know about chnages in output latency.

This change is related to streaming video player audio through the
Yousician audio engine, and the need to know the actual latency there.

YS-9374